+++ unnamed_engine/src/core/mod.rs
@@ -1,3 +1,4 @@
 //! ## Core
+pub mod module;
 pub mod state;
 pub mod engine;