+++ unnamed_engine/src/core/state.rs
@@ -1,38 +1,91 @@
 //! ## State
 //!
 //! Runtime data.
-use std::{sync::Arc, time::Duration};
+use std::{error::Error, sync::Arc};
 
-use winit::window::Window;
+use winit::{event::WindowEvent, window::Window};
 
-use crate::{event::event::Event, renderer::renderer::Renderer};
+use crate::{event::event::Event, renderer::{renderer::Renderer, viewport::Viewport}};
+
+use super::module::Module;
 
 pub struct State {
-  pub renderer: Renderer,
+  pub modules: Vec<Box<dyn Module>>,
 }
 
 impl State {
   pub async fn new(viewport: (Arc<Window>, wgpu::Color)) -> Self {
-    let renderer = Renderer::new(viewport).await;
+    let mut modules: Vec<Box<dyn Module>> = Vec::new();
+
+    let renderer_module = Box::new(Renderer::new(viewport).await);
+
+    modules.push(renderer_module);
 
     Self {
-      renderer,
+      modules,
     }
   }
 
-  pub fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>) {
-    self.renderer.resize(new_size);
+  pub fn register_module(&mut self, module: Box<dyn Module>) {
+    self.modules.push(module);
   }
 
-  pub fn process_events(&mut self, event: Event) -> bool {
-    self.renderer.process_events(event)
+  pub fn process_events(&mut self, event: Event) -> Result<(), Box<dyn Error>> {
+    for module in &mut self.modules {
+      match module.process_events(event) {
+        Ok(_) => {},
+        Err(e) => {
+          log::error!("Failed to process events: {e}");
+          return Err(e);
+        }
+      }
+    }
+
+    Ok(())
   }
 
-  pub fn update(&mut self, dt: Duration) {
-    self.renderer.update(dt);
+  pub fn update(&mut self, dt: instant::Duration) -> Result<(), Box<dyn Error>> {
+    for module in &mut self.modules {
+      match module.update(dt) {
+        Ok(_) => {},
+        Err(e) => {
+          log::error!("Failed to update: {e}");
+          return Err(e);
+        }
+      }
+    }
+
+    Ok(())
   }
 
-  pub fn render(&mut self) -> Result<(), wgpu::SurfaceError> {
-    self.renderer.render()
+  pub fn render(&mut self) -> Result<(), Box<dyn Error>> {
+    for module in &mut self.modules {
+      match module.render() {
+        Ok(_) => {},
+        Err(e) => {
+          log::error!("Failed to render: {e}");
+          return Err(e);
+        }
+      }
+    }
+
+    Ok(())
+  }
+
+  pub fn find_viewport(&mut self) -> Option<&mut Viewport> {
+    for module in &mut self.modules {
+      match module.viewport() {
+        Some(viewport) => return Some(viewport),
+        None => return None,
+      }
+    }
+
+    None
+  }
+
+  pub fn forward_window_event(&mut self, event: &WindowEvent) {
+    for module in &mut self.modules {
+      module.process_window_events(event);
+    }
   }
 }