+++ unnamed_engine/src/gui/egui_renderer.rs
@@ -66,13 +66,13 @@ impl EguiRenderer {
     screen_descriptor: ScreenDescriptor,
     run_ui: impl FnOnce(&Context),
   ) {
-    let raw_input = self.state.take_egui_input(&window);
+    let raw_input = self.state.take_egui_input(window);
     let full_output = self.context.run(raw_input, |_ui| {
       run_ui(&self.context);
     });
 
     self.state
-      .handle_platform_output(&window, full_output.platform_output);
+      .handle_platform_output(window, full_output.platform_output);
 
     let tris = self
       .context
@@ -80,15 +80,15 @@ impl EguiRenderer {
 
     for (id, image_delta) in &full_output.textures_delta.set {
       self.renderer
-        .update_texture(&device, &queue, *id, &image_delta);
+        .update_texture(device, queue, *id, image_delta);
     }
 
     self.renderer
-      .update_buffers(&device, &queue, encoder, &tris, &screen_descriptor);
+      .update_buffers(device, queue, encoder, &tris, &screen_descriptor);
 
     let mut rpass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
       color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-        view: &window_surface_view,
+        view: window_surface_view,
         resolve_target: None,
         ops: wgpu::Operations {
           load: wgpu::LoadOp::Load,