+++ unnamed_engine/src/input/keyboard.rs
@@ -48,3 +48,9 @@ impl Keyboard {
     }
   }
 }
+
+impl Default for Keyboard {
+    fn default() -> Self {
+      Self::new()
+    }
+}